Found better early game start patch methods for both with and without Fika, fixes issue four for HomeComforts
