+++ LeaveItThere-Core/Patches/EarlyGameStartedPatch.cs
@@ -1,0 +1,43 @@
+﻿using EFT;
+using HarmonyLib;
+using LeaveItThere.Components;
+using LeaveItThere.Fika;
+using SPT.Reflection.Patching;
+using System.Reflection;
+
+namespace LeaveItThere.Patches
+{
+    // This patch is only enabled when Fika is installed
+    internal class EarlyGameStartedPatchFika : ModulePatch
+    {
+        protected override MethodBase GetTargetMethod()
+        {
+            return AccessTools.Method(typeof(GameWorld), nameof(GameWorld.RegisterRestrictableZones));
+        }
+
+        [PatchPostfix]
+        static void PatchPrefix()
+        {
+            Plugin.DebugLog("happened");
+            LITSession.CreateNewModSession();
+            ObjectMover.CreateNewObjectMover();
+        }
+    }
+
+    // This patch is only enabled when Fika is NOT installed
+    internal class EarlyGameStartedPatch : ModulePatch
+    {
+        protected override MethodBase GetTargetMethod()
+        {
+            return AccessTools.Method(typeof(BotsController), nameof(BotsController.SetSettings));
+        }
+
+        [PatchPostfix]
+        static void PatchPrefix()
+        {
+            Plugin.DebugLog("happened");
+            LITSession.CreateNewModSession();
+            ObjectMover.CreateNewObjectMover();
+        }
+    }
+}
